fix(groups): correct static page account creation flow
